+++ Sources/KlaviyoSwift/Klaviyo.swift
@@ -12,116 +12,7 @@ import UIKit
 let version = "2.0.0"
 
 public class Klaviyo: NSObject  {
-    
-    public struct Event {
-        public struct Attributes {
-            public struct Metric {
-                public let name: String
-                public let service: String?
-                public init(name: String,
-                            service: String?="ios-analytics") {
-                    self.name = name
-                    self.service = service
-                }
-            }
-            public let metric: Metric
-            public let properties: [String: Any]
-            public let profile: [String: Any]
-            public var time: Date?
-            public let value: Double?
-            public let uniqueId: String
-            public init(metric: Metric,
-                        properties: [String : Any],
-                        profile: [String : Any],
-                        value: Double? = nil,
-                        time: Date? = nil,
-                        uniqueId: String? = nil) {
-                self.profile = profile
-                self.metric = metric
-                self.properties = properties
-                self.value = value
-                self.time = time
-                self.time = time ?? environment.date()
-                self.uniqueId = uniqueId ?? environment.uuid().uuidString
-            }
-            
-        }
-        public let attributes: Attributes
-        public init(attributes: Attributes) {
-            self.attributes = attributes
-        }
-    }
-    
-    public struct Profile {
-        public struct Attributes {
-            public struct Location {
-                public let address1: String?
-                public let address2: String?
-                public let city: String?
-                public let country: String?
-                public let latitude: Double?
-                public let longitude: Double?
-                public let region: String?
-                public let zip: String?
-                public let timezone: String?
-                public init(address1: String?=nil,
-                            address2: String?=nil,
-                            city: String?=nil,
-                            country: String?=nil,
-                            latitude: Double?=nil,
-                            longitude: Double?=nil,
-                            region: String?=nil,
-                            zip: String?=nil,
-                            timezone: String?=TimeZone.autoupdatingCurrent.identifier) {
-                    self.address1 = address1
-                    self.address2 = address2
-                    self.city = city
-                    self.country = country
-                    self.latitude = latitude
-                    self.longitude = longitude
-                    self.region = region
-                    self.zip = zip
-                    self.timezone = timezone
-                }
-            }
-            let email: String?
-            let phoneNumber: String?
-            let externalId: String?
-            let firstName: String?
-            let lastName: String?
-            let organization: String?
-            let title: String?
-            let image: String?
-            let location: Location?
-            let properties: [String: Any]?
-            public init(email: String?=nil,
-                 phoneNumber: String?=nil,
-                 externalId: String?=nil,
-                 firstName: String?=nil,
-                 lastName: String?=nil,
-                 organization: String?=nil,
-                 title: String?=nil,
-                 image: String?=nil,
-                 location: Location?=nil,
-                 properties: [String : Any]?=nil) {
-                self.email = email
-                self.phoneNumber = phoneNumber
-                self.externalId = externalId
-                self.firstName = firstName
-                self.lastName = lastName
-                self.organization = organization
-                self.title = title
-                self.image = image
-                self.location = location
-                self.properties = properties
-            }
-        }
-        let attributes: Attributes
-        public init(attributes: Attributes) {
-            self.attributes = attributes
-        }
-    }
-    
+   
     /*
     Klaviyo Class Constants
     */
